requirements.txt
pillow
customtkinter
CTkMessagebox
matplotlib
seaborn
numpy
pandas